Improve ayo quest 1.2.8 integrando com pg railws e retirando supabase

+++ src/main/java/br/com/ayo_quest/ayo_quest/dto/proficiencia/CadastrarProficienciaDTO.java
@@ -20,7 +20,7 @@ public class CadastrarProficienciaDTO {
 
     private String descricao;
 
-    private Integer notaMinima;
+    private Long notaMinima;
 
     private Integer totalQuestao;
